# Optimise xorBytes - huge gains for SRTP AES-CM