```
name                           old time/op    new time/op    delta
EncryptRTP/CTR-100-12             852ns ± 1%     758ns ± 2%  -11.09%  (p=0.000 n=9+10)
EncryptRTP/CTR-1000-12           3.98µs ± 1%    2.96µs ± 2%  -25.69%  (p=0.000 n=10+10)
EncryptRTPInPlace/CTR-100-12      824ns ± 1%     734ns ± 4%  -10.93%  (p=0.000 n=10+9)
EncryptRTPInPlace/CTR-1000-12    3.89µs ± 5%    2.85µs ± 1%  -26.77%  (p=0.000 n=10+10)
DecryptRTP/CTR-12                 518ns ± 1%     515ns ± 1%   -0.70%  (p=0.040 n=9+9)
Write/CTR-100-12                  863ns ± 1%     762ns ± 1%  -11.71%  (p=0.000 n=10+9)
Write/CTR-1000-12                3.91µs ± 2%    2.90µs ± 2%  -25.85%  (p=0.000 n=9+9)
WriteRTP/CTR-100-12               816ns ± 4%     694ns ± 2%  -14.93%  (p=0.000 n=10+10)
WriteRTP/CTR-1400-12             3.95µs ± 3%    2.79µs ± 1%  -29.21%  (p=0.000 n=10+9)

name                           old speed      new speed      delta
EncryptRTP/CTR-100-12           131MB/s ± 1%   148MB/s ± 2%  +12.48%  (p=0.000 n=9+10)
EncryptRTP/CTR-1000-12          254MB/s ± 1%   342MB/s ± 2%  +34.58%  (p=0.000 n=10+10)
EncryptRTPInPlace/CTR-100-12    136MB/s ± 1%   153MB/s ± 3%  +12.29%  (p=0.000 n=10+9)
EncryptRTPInPlace/CTR-1000-12   260MB/s ± 5%   355MB/s ± 1%  +36.51%  (p=0.000 n=10+10)
DecryptRTP/CTR-12              54.0MB/s ± 1%  54.4MB/s ± 1%   +0.70%  (p=0.040 n=9+9)
Write/CTR-100-12                130MB/s ± 1%   147MB/s ± 1%  +13.25%  (p=0.000 n=10+9)
Write/CTR-1000-12               259MB/s ± 2%   349MB/s ± 2%  +34.85%  (p=0.000 n=9+9)
WriteRTP/CTR-100-12             137MB/s ± 4%   161MB/s ± 2%  +17.52%  (p=0.000 n=10+10)
WriteRTP/CTR-1400-12            256MB/s ± 3%   362MB/s ± 1%  +41.22%  (p=0.000 n=10+9)
```